Split FW-4 test coverage
